Add files via upload
The constraint here is to distribute apples and oranges equally among the contestants and at the same time, every fruit with Rushitote must be distributed. so we can compute gcd of no of apples and oranges. The while loop here computes the gcd of apples and oranges.

procedure to compute gcd:

1)if n=0 return the value of m else step 2.
2)divide m by n and assign the remainder to r.
3)Assign the value of n to m and the value of r to n.